Added: .gitignore Changed: Moved template functions to separate file such that they can be reused.
